+++ packages_rs/nextclade-cli/src/cli/nextalign_cli.rs
@@ -1,13 +1,12 @@
-use clap::{AppSettings, ArgEnum, ArgGroup, CommandFactory, Parser, Subcommand, ValueHint};
+use crate::cli::common::get_fasta_basename;
+use crate::cli::verbosity::{Verbosity, WarnLevel};
+use clap::{AppSettings, ArgEnum, CommandFactory, Parser, Subcommand, ValueHint};
 use clap_complete::{generate, Generator, Shell};
 use clap_complete_fig::Fig;
-use clap_verbosity_flag::{Verbosity, WarnLevel};
 use eyre::{eyre, ContextCompat, Report, WrapErr};
 use itertools::Itertools;
 use lazy_static::lazy_static;
-use log::LevelFilter;
 use nextclade::align::params::AlignPairwiseParamsOptional;
-use nextclade::io::fs::basename;
 use nextclade::make_error;
 use nextclade::utils::global_init::setup_logger;
 use std::fmt::Debug;
@@ -18,7 +17,6 @@ use strum_macros::EnumIter;
 
 lazy_static! {
   static ref SHELLS: &'static [&'static str] = &["bash", "elvish", "fish", "fig", "powershell", "zsh"];
-  static ref VERBOSITIES: &'static [&'static str] = &["off", "error", "warn", "info", "debug", "trace"];
 }
 
 #[derive(Parser, Debug)]
@@ -37,17 +35,9 @@ pub struct NextalignArgs {
   #[clap(subcommand)]
   pub command: NextalignCommands,
 
-  /// Set verbosity level [default: warn]
-  #[clap(long, global = true, conflicts_with = "verbose", conflicts_with = "silent", possible_values(VERBOSITIES.iter()))]
-  pub verbosity: Option<LevelFilter>,
-
-  /// Disable all console output. Same as --verbosity=off
-  #[clap(long, global = true, conflicts_with = "verbose", conflicts_with = "verbosity")]
-  pub silent: bool,
-
   /// Make output more quiet or more verbose
-  #[clap(flatten)]
-  pub verbose: Verbosity<WarnLevel>,
+  #[clap(flatten, next_help_heading = "  Verbosity")]
+  pub verbosity: Verbosity<WarnLevel>,
 }
 
 #[derive(Subcommand, Debug)]
@@ -81,11 +71,27 @@ pub enum NextalignOutputSelection {
 }
 
 #[derive(Parser, Debug)]
-pub struct NextalignRunArgs {
-  /// Path to a FASTA file with input sequences
+pub struct NextalignRunInputArgs {
+  /// Path to one or multiple FASTA files with input sequences
+  ///
+  /// Accepts plain or compressed FASTA files. If a compressed fasta file is provided, it will be transparently
+  /// decompressed. Supported compression formats: `gz`, `bz2`, `xz`, `zstd`. Decompressor is chosen based on file
+  /// extension. If there's multiple input files, then different files can have different compression formats.
+  ///
+  /// If no input files provided, the plain fasta input is read from standard input (stdin).
+  ///
+  /// See: https://en.wikipedia.org/wiki/FASTA_format
+  #[clap(value_hint = ValueHint::FilePath)]
+  #[clap(display_order = 1)]
+  pub input_fastas: Vec<PathBuf>,
+
+  /// REMOVED. Use positional arguments instead.
+  ///
+  /// Example: nextalign run -D dataset/ -O out/ seq1.fasta seq2.fasta
   #[clap(long, short = 'i', visible_alias("sequences"))]
   #[clap(value_hint = ValueHint::FilePath)]
-  pub input_fasta: PathBuf,
+  #[clap(hide_long_help = true, hide_short_help = true)]
+  pub input_fasta: Option<PathBuf>,
 
   /// Path to a FASTA file containing reference sequence.
   ///
@@ -124,14 +130,22 @@ pub struct NextalignRunArgs {
   )]
   #[clap(value_hint = ValueHint::FilePath)]
   pub genes: Option<Vec<String>>,
+}
+
+#[derive(Parser, Debug)]
+pub struct NextalignRunOutputArgs {
+  /// REMOVED. Use `--output-all` instead
+  #[clap(long)]
+  #[clap(value_hint = ValueHint::DirPath)]
+  #[clap(hide_long_help = true, hide_short_help = true)]
+  pub output_dir: Option<PathBuf>,
 
   /// Produce all of the output files into this directory, using default basename and predefined suffixes and extensions. This is equivalent to specifying each of the individual `--output-*` flags. Convenient when you want to receive all or most of output files into the same directory and don't care about their filenames.
   ///
   /// Output files can be optionally included or excluded using `--output-selection` flag.
   /// The base filename can be set using `--output-basename` flag.
   ///
-  /// If both the `--output-all` and individual `--output-*` flags are provided, each
-  //  individual flag overrides the corresponding default output path.
+  /// If both the `--output-all` and individual `--output-*` flags are provided, each individual flag overrides the corresponding default output path.
   ///
   /// At least one of the output flags is required: `--output-all`, `--output-fasta`, `--output-translations`, `--output-insertions`, `--output-errors`
   ///
@@ -171,6 +185,9 @@ pub struct NextalignRunArgs {
   ///
   /// Takes precedence over paths configured with `--output-all`, `--output-basename` and `--output-selection`.
   ///
+  /// If filename ends with one of the supported file extensions: `gz`, `bz2`, `xz`, `zstd`, it will be transparently
+  /// compressed. If a filename is "-" then the output will be written uncompressed to standard output (stdout).
+  ///
   /// If the required directory tree does not exist, it will be created.
   #[clap(long, short = 'o')]
   #[clap(value_hint = ValueHint::AnyPath)]
@@ -181,6 +198,9 @@ pub struct NextalignRunArgs {
   /// Make sure you properly quote and/or escape the curly braces, so that your shell, programming language or pipeline manager does not attempt to substitute the variables.
   ///
   /// Takes precedence over paths configured with `--output-all`, `--output-basename` and `--output-selection`.
+  ///
+  /// If filename ends with one of the supported file extensions: `gz`, `bz2`, `xz`, `zstd`, it will be transparently
+  /// compressed.
   ///
   /// Example for bash shell:
   ///
@@ -195,6 +215,9 @@ pub struct NextalignRunArgs {
   ///
   /// Takes precedence over paths configured with `--output-all`, `--output-basename` and `--output-selection`.
   ///
+  /// If filename ends with one of the supported file extensions: `gz`, `bz2`, `xz`, `zstd`, it will be transparently
+  /// compressed. If a filename is "-" then the output will be written uncompressed to standard output (stdout).
+  ///
   /// If the required directory tree does not exist, it will be created.
   #[clap(long, short = 'I')]
   #[clap(value_hint = ValueHint::AnyPath)]
@@ -203,6 +226,9 @@ pub struct NextalignRunArgs {
   /// Path to output CSV file containing errors and warnings occurred during processing
   ///
   /// Takes precedence over paths configured with `--output-all`, `--output-basename` and `--output-selection`.
+  ///
+  /// If filename ends with one of the supported file extensions: `gz`, `bz2`, `xz`, `zstd`, it will be transparently
+  /// compressed. If a filename is "-" then the output will be written uncompressed to standard output (stdout).
   ///
   /// If the required directory tree does not exist, it will be created.
   #[clap(long, short = 'e')]
@@ -224,12 +250,37 @@ pub struct NextalignRunArgs {
   #[clap(long)]
   pub in_order: bool,
 
-  /// Number of processing jobs. If not specified, all available CPU threads will be used.
-  #[clap(global = false, long, short = 'j', default_value_t = num_cpus::get() )]
-  pub jobs: usize,
+  /// Replace unknown nucleotide characters with 'N'
+  ///
+  /// By default, the sequences containing unknown nucleotide nucleotide characters are skipped with a warning - they
+  /// are not aligned and not included into results. If this flag is provided, then before the alignment,
+  /// all unknown characters are replaced with 'N'. This replacement allows to align these sequences.
+  ///
+  /// The following characters are considered known:  '-', 'A', 'B', 'C', 'D', 'G', 'H', 'K', 'M', 'N', 'R', 'S', 'T', 'V', 'W', 'Y'
+  #[clap(long)]
+  pub replace_unknown: bool,
+}
 
-  #[clap(flatten)]
+#[derive(Parser, Debug)]
+pub struct NextalignRunOtherArgs {
+  /// Number of processing jobs. If not specified, all available CPU threads will be used.
+  #[clap(global = false, long, short = 'j', default_value_t = num_cpus::get())]
+  pub jobs: usize,
+}
+
+#[derive(Parser, Debug)]
+pub struct NextalignRunArgs {
+  #[clap(flatten, next_help_heading = "  Inputs")]
+  pub inputs: NextalignRunInputArgs,
+
+  #[clap(flatten, next_help_heading = "  Outputs")]
+  pub outputs: NextalignRunOutputArgs,
+
+  #[clap(flatten, next_help_heading = "  Alignment parameters")]
   pub alignment_params: AlignPairwiseParamsOptional,
+
+  #[clap(flatten, next_help_heading = "  Other")]
+  pub other: NextalignRunOtherArgs,
 }
 
 fn generate_completions(shell: &str) -> Result<(), Report> {
@@ -253,22 +304,39 @@ fn generate_completions(shell: &str) -> Result<(), Report> {
 /// Get output filenames provided by user or, if not provided, create filenames based on input fasta
 pub fn nextalign_get_output_filenames(run_args: &mut NextalignRunArgs) -> Result<(), Report> {
   let NextalignRunArgs {
-    input_fasta,
-    output_all,
-    ref mut output_basename,
-    ref mut output_errors,
-    ref mut output_fasta,
-    ref mut output_insertions,
-    ref mut output_translations,
-    ref mut output_selection,
-    ..
+    inputs:
+      NextalignRunInputArgs {
+        input_fastas,
+        input_ref,
+        input_gene_map,
+        genes,
+        ..
+      },
+    outputs:
+      NextalignRunOutputArgs {
+        output_all,
+        output_basename,
+        output_selection,
+        output_fasta,
+        output_translations,
+        output_insertions,
+        output_errors,
+        include_reference,
+        in_order,
+        ..
+      },
+    other: NextalignRunOtherArgs { jobs },
+    alignment_params,
   } = run_args;
 
   // If `--output-all` is provided, then we need to deduce default output filenames,
   // while taking care to preserve values of any individual `--output-*` flags,
   // as well as to honor restrictions put by the `--output-selection` flag, if provided.
   if let Some(output_all) = output_all {
-    let output_basename = output_basename.get_or_insert(basename(&input_fasta)?);
+    let output_basename = output_basename
+      .clone()
+      .unwrap_or_else(|| get_fasta_basename(input_fastas).unwrap_or_else(|| "nextalign".to_owned()));
+
     let default_output_file_path = output_all.join(&output_basename);
 
     // If `--output-selection` is empty or contains `all`, then fill it with all possible variants
@@ -347,26 +415,58 @@ At least one of the following flags is required:
   Ok(())
 }
 
+const ERROR_MSG_INPUT_FASTA_REMOVED: &str = r#"The argument `--input-fasta` (alias: `--sequences`, `-i`) is removed in favor of positional arguments.
+
+Try:
+
+  nextalign run -r ref.fasta -m genemap.gff -O out/ seq1.fasta seq2.fasta
+
+                                                       ^          ^
+                                              one or multiple positional arguments
+                                                with paths to input fasta files
+
+
+When positional arguments are not provided, nextalign will read input fasta from standard input.
+
+For more information, type
+
+  nextalign run --help"#;
+
+const ERROR_MSG_OUTPUT_DIR_REMOVED: &str = r#"The argument `--output-dir` is removed in favor of `--output-all`.
+
+When provided, `--output-all` allows to write all possible outputs into a directory.
+
+The defaut base name of the files can be overriden with `--output-basename` argument.
+
+The set of output files can be restricted with `--output-selection` argument.
+
+For more information, type:
+
+  nextalign run --help"#;
+
+pub fn nextalign_check_removed_args(run_args: &mut NextalignRunArgs) -> Result<(), Report> {
+  if run_args.inputs.input_fasta.is_some() {
+    return make_error!("{ERROR_MSG_INPUT_FASTA_REMOVED}");
+  }
+
+  if run_args.outputs.output_dir.is_some() {
+    return make_error!("{ERROR_MSG_OUTPUT_DIR_REMOVED}");
+  }
+
+  Ok(())
+}
+
 pub fn nextalign_parse_cli_args() -> Result<NextalignArgs, Report> {
   let mut args = NextalignArgs::parse();
 
-  // --verbosity=<level> and --silent take priority over -v and -q
-  let filter_level = if args.silent {
-    LevelFilter::Off
-  } else {
-    match args.verbosity {
-      None => args.verbose.log_level_filter(),
-      Some(verbosity) => verbosity,
-    }
-  };
-
-  setup_logger(filter_level);
+  setup_logger(args.verbosity.get_filter_level());
 
   match &mut args.command {
     NextalignCommands::Completions { shell } => {
       generate_completions(shell).wrap_err_with(|| format!("When generating completions for shell '{shell}'"))?;
     }
     NextalignCommands::Run(ref mut run_args) => {
+      nextalign_check_removed_args(run_args)?;
       nextalign_get_output_filenames(run_args).wrap_err("When deducing output filenames")?;
     }
   }